Push finishing tubes

+++ kuliah/pertemuan14/index.php
@@ -32,36 +32,40 @@ if (isset($_POST["cari"])) {
 
     </form>
     <br>
-    <table border="1" cellpadding="10" cellspacing="0">
+    <?php if ($mahasiswa) : ?>
+        <table border="1" cellpadding="10" cellspacing="0">
 
-        <tr>
-            <th>No.</th>
-            <th>Aksi</th>
-            <th>Gambar</th>
-            <th>NRP</th>
-            <th>Nama</th>
-            <th>Email</th>
-            <th>Jurusan</th>
-        </tr>
-        <?php $i = 1; ?>
-        <?php foreach ($mahasiswa as $row) : ?>
             <tr>
-                <td><?= $i; ?></td>
-                <td>
-                    <a href="ubah.php?id=<?= $row["id"]; ?>">ubah</a> |
-                    <a href="hapus.php?id=<?= $row["id"]; ?>" onclick="return confirm('yakin?')">hapus</a>
-                </td>
-                <td><img src="img/<?= $row["gambar"]; ?>" width="50" alt=""></td>
-                <td><?= $row["nrp"]; ?></td>
-                <td><?= $row["nama"]; ?></td>
-                <td><?= $row["email"]; ?></td>
-                <td><?= $row["jurusan"]; ?></td>
-
+                <th>No.</th>
+                <th>Aksi</th>
+                <th>Gambar</th>
+                <th>NRP</th>
+                <th>Nama</th>
+                <th>Email</th>
+                <th>Jurusan</th>
             </tr>
-            <?php $i++; ?>
-        <?php endforeach; ?>
+            <?php $i = 1; ?>
+            <?php foreach ($mahasiswa as $row) : ?>
+                <tr>
+                    <td><?= $i; ?></td>
+                    <td>
+                        <a href="ubah.php?id=<?= $row["id"]; ?>">ubah</a> |
+                        <a href="hapus.php?id=<?= $row["id"]; ?>" onclick="return confirm('yakin?')">hapus</a>
+                    </td>
+                    <td><img src="img/<?= $row["gambar"]; ?>" width="50" alt=""></td>
+                    <td><?= $row["nrp"]; ?></td>
+                    <td><?= $row["nama"]; ?></td>
+                    <td><?= $row["email"]; ?></td>
+                    <td><?= $row["jurusan"]; ?></td>
 
-    </table>
+                </tr>
+                <?php $i++; ?>
+            <?php endforeach; ?>
+
+        </table>
+    <?php else : ?>
+        <h1>Data Tidak Tersedia</h1>
+    <?php endif; ?>
 
 </body>
